Proper handling of unit changes for vintrange and vintall.
For whatever reason, the Fred API does not always handle a change of
units well when calling vintall or vintrange with a "units" option that
is not just "lin". Often, nothing gets returned.

This commit fixes that. The "units" key (and its value) are removed from
the varargin lists that would otherwise be passed to the Fred API call.
Then, at the end, we use the value in a call to transform.m.