Add reshaping into array structures by vintage date.
toDataset flag has been modified to toDatasetByVintage flag. This
signifies that the data structures returned by fred.vint and fred.latest
should be array structs where each entry is a different vintage date and
the "value" field is a data matrix merging multiple series with possibly
mixed frequencies.

This used to happen for calls to fred.vint (with one vintage) and
fred.latest. Now we can do it for calls with multiple vintage dates.