Fixed bug in dtnum that breaks for ver<2014a
fred.dtnum takes as input a string or cell array of strings and returns
Matlab datenums. Given these inputs, there is no requirement that the
input strings conform to the format 'yyyy-mm-dd'. This was breaking for
versions <2014a. For 2014a, the call to fred.dtnum(date()) was parsing a
date that is by default in format dd-mmm-yyyy as the above format,
returning a date in 2035.