Add fred_fmt optional argument for fred.dtnum
Now fred.dtnum can accept any arbitrary date string format, but we can
also specify that it is in the fred format (yyyy-mm-dd) as an optional
argument if we would like.

This allows us to specify that it is in fred format behind the scenes
(so we get speedups, since datenum without a format string is slow), but
still let the user pass any datestr format to be converted.